Move date getter/setter to mio settings service
