+++ libs/ts/movein/ngx/src/lib/modules/builder/modules/zone-select/zone-select.component.ts
@@ -6,6 +6,7 @@ import { Angulartics2 } from 'angulartics2';
 import { LocalStoreService } from '@tamu-gisc/common/ngx/local-store';
 
 import { ResidenceHall, ResidenceZones } from '../../../../interfaces/move-in-out.interface';
+import { RESIDENCES } from '../../../../dictionaries/move-in-out.dictionary';
 
 @Component({
   selector: 'tamu-gisc-zone-select',
@@ -16,131 +17,7 @@ export class ZoneSelectComponent implements OnInit {
   public savedResidence: ResidenceHall;
   private halls = [];
 
-  public zones: ResidenceZones = {
-    southSide: {
-      name: 'South Side',
-      halls: [
-        {
-          name: 'Appelt Hall',
-          Bldg_Number: ['0293']
-        },
-        {
-          name: 'Aston Hall',
-          Bldg_Number: ['0447']
-        },
-        {
-          name: 'Dun Hall',
-          Bldg_Number: ['0442']
-        },
-        {
-          name: 'Eppright Hall',
-          Bldg_Number: ['0292']
-        },
-        {
-          name: 'Hart Hall',
-          Bldg_Number: ['0417']
-        },
-        {
-          name: 'Kreuger Hall',
-          Bldg_Number: ['0441']
-        },
-        {
-          name: 'Mosher Hall',
-          Bldg_Number: ['0433']
-        },
-        {
-          name: 'Rudder Hall',
-          Bldg_Number: ['0291']
-        },
-        {
-          name: 'Underwood Hall',
-          Bldg_Number: ['0394']
-        },
-        {
-          name: 'Wells Hall',
-          Bldg_Number: ['0290']
-        }
-      ]
-    },
-    northSide: {
-      name: 'North Side',
-      halls: [
-        {
-          name: 'Clements Hall',
-          Bldg_Number: ['0548']
-        },
-        {
-          name: 'Davis-Gary Hall',
-          Bldg_Number: ['0415']
-        },
-        {
-          name: 'Fowler Hall',
-          Bldg_Number: ['0427']
-        },
-        {
-          name: 'Haas Hall',
-          Bldg_Number: ['0549']
-        },
-        {
-          name: 'Hart Hall ',
-          Bldg_Number: ['0417']
-        },
-        {
-          name: 'Hobby Hall',
-          Bldg_Number: ['0653']
-        },
-        {
-          name: 'Hughes Hall',
-          Bldg_Number: ['0426']
-        },
-        {
-          name: 'Hullabaloo Hall',
-          Bldg_Number: ['1416']
-        },
-        {
-          name: 'Keathley Hall',
-          Bldg_Number: ['0428']
-        },
-        {
-          name: 'Lechner Hall',
-          Bldg_Number: ['0294']
-        },
-        {
-          name: 'Legett Hall',
-          Bldg_Number: ['0419']
-        },
-        {
-          name: 'McFadden Hall',
-          Bldg_Number: ['0550']
-        },
-        {
-          name: 'Moses Hall',
-          Bldg_Number: ['0412']
-        },
-        {
-          name: 'Neeley Hall',
-          Bldg_Number: ['0652']
-        },
-        {
-          name: 'Schuhmacher Hall',
-          Bldg_Number: ['0430']
-        },
-        {
-          name: 'Walton Hall',
-          Bldg_Number: ['0430']
-        }
-      ]
-    },
-    whiteCreek: {
-      name: 'White Creek',
-      halls: [
-        {
-          name: 'White Creek Apartments',
-          Bldg_Number: ['1590', '1591', '1592', '0064']
-        }
-      ]
-    }
-  };
+  public zones: ResidenceZones = RESIDENCES;
 
   constructor(
     private store: LocalStoreService,
